Resolve logs authentication issue on RHEL OpenShift
Related to https://github.com/kubernetes/kubernetes/issues/63128

According to the CRC documentation:

> The system bundle in each released crc executable expires 30 days after the release.

Although unverified, it's likely that the certificates were not being updated
automatically giving us an obscure authentication error. Changing the crc version
to "latest" resolves the issue. We also needed to make additional changes to address
a new error on the latest version related to:

https://access.redhat.com/solutions/4661741

Signed-off-by: Javier Romero <rjavier@vmware.com>